Externalised sendEmailVerification
