+++ instafood_frontend/src/authentication/signup.js
@@ -2,7 +2,8 @@ import { useNavigate, Link } from "react-router-dom";
 import { useState, useEffect } from "react";
 
 import { auth, db, functions } from "../firebaseConf";
-import { createUserWithEmailAndPassword, sendEmailVerification, signOut } from "firebase/auth";
+import { createUserWithEmailAndPassword, signOut } from "firebase/auth";
+import SendEmailVerification from "./sendEmailVerification";
 import { doc, getDoc } from 'firebase/firestore';
 
 import { httpsCallable } from 'firebase/functions';
@@ -26,7 +27,9 @@ function Signup() {
         async function checkUserIDUnique() {
             const snapshot = await getDoc(userIDsRef);
             const userIDs = snapshot.data().userIDs;
-            setUserIDUnique(!userIDs.includes(userID));
+            if (!userIDs.includes(userID) && userID !== "") {
+                setUserIDUnique(true);
+            }
         }
         checkUserIDUnique();
     }, [userID]);
@@ -35,7 +38,7 @@ function Signup() {
         e.preventDefault();
 
         createUserWithEmailAndPassword(auth, email, password)
-            .then(() => {
+            .then( async () => {
                 const user = auth.currentUser;
 
                 createUserProfile({
@@ -47,26 +50,9 @@ function Signup() {
                 })
 
                 // Send email verification
-                // Direct users back to app and sign them in when they click on the link in the email
-                const continueUrl = "http://localhost:3000/signInAfterEmailVerification";
-                const params = new URLSearchParams();
-                params.append("userEmail", email);
-                params.append("password", password);
-
-                const actionCodeSettings = {
-                    url: continueUrl + "?" + params.toString(),
-                    handleCodeInApp: true,
-                };
-
-                sendEmailVerification(user, actionCodeSettings)
-                    .then(() => {
-                        alert("Email verification sent. Please check your email and follow the instructions to verify your email address.");
-                        signOut(auth);
-                        navigate("/");
-                    })
-                    .catch((error) => {
-                        console.log(error);
-                    });
+                await SendEmailVerification(email, password);
+                signOut(auth);
+                navigate("/");
             })
 
             .catch((error) => {
@@ -82,79 +68,89 @@ function Signup() {
                     alert("Password must be at least 6 characters.");
                 }
 
+                // missing password
+                if (error.code === "auth/missing-password") {
+                    alert("Please enter a password.");
+                }
+
                 navigate("/signup");
             });
     }
 
     return (
         <div>
-            <div>
-                <label>Email</label>
-                <input
-                    type="email"
-                    value={email}
-                    onChange={(e) => setEmail(e.target.value)}
-                />
-            </div>
+            <form onSubmit={handleSignup}>
+                <div>
+                    <label>Email</label>
+                    <input
+                        type="email"
+                        placeholder="Required"
+                        value={email}
+                        onChange={(e) => setEmail(e.target.value)}
+                    />
+                </div>
 
-            <div>
-                <label>Password</label>
-                <input
-                    type="password"
-                    value={password}
-                    onChange={(e) => setPassword(e.target.value)}
-                />
-            </div>
+                <div>
+                    <label>Password</label>
+                    <input
+                        type="password"
+                        placeholder="Required"
+                        value={password}
+                        onChange={(e) => setPassword(e.target.value)}
+                    />
+                </div>
 
-            <div>
-                <label>Username</label>
-                <input
-                    type="text"
-                    required
-                    value={username}
-                    onChange={(e) => setUserName(e.target.value)}
-                />
-            </div>
+                <div>
+                    <label>Username</label>
+                    <input
+                        type="text"
+                        placeholder="Required"
+                        required
+                        value={username}
+                        onChange={(e) => setUserName(e.target.value)}
+                    />
+                </div>
 
-            <div>
-                <label>Bio</label>
-                <input
-                    type="text"
-                    value={bio}
-                    onChange={(e) => setBio(e.target.value)}
-                />
-            </div>
+                <div>
+                    <label>Bio</label>
+                    <input
+                        type="text"
+                        value={bio}
+                        onChange={(e) => setBio(e.target.value)}
+                    />
+                </div>
 
-            <div>
-                <p>UserID cannot be changed later</p>
-                <label>User ID</label>
-                <input
-                    type="text"
-                    required
-                    value={userID}
-                    onChange={(e) => setUserID(e.target.value)}
-                />
-            </div>
+                <div>
+                    <p>UserID cannot be changed later</p>
+                    <label>User ID</label>
+                    <input
+                        type="text"
+                        placeholder="Required"
+                        required
+                        value={userID}
+                        onChange={(e) => setUserID(e.target.value)}
+                    />
+                </div>
 
-            <div>
-                <label>Set as private</label>
-                <input
-                    type="checkbox"
-                    required
-                    checked={isPrivate}
-                    onChange={(e) => setIsPrivate(e.target.checked)}
-                />
-            </div>
+                <div>
+                    <label>Set as private</label>
+                    <input
+                        type="checkbox"
+                        checked={isPrivate}
+                        onChange={(e) => setIsPrivate(e.target.checked)}
+                    />
+                </div>
 
-            {
-                userIDUnique ? (
-                    <div>
-                        <p>User ID is unique!</p>
-                        <button onClick={handleSignup}>Sign Up</button>
-                    </div>
-                )
-                    : (<p>User ID already in use. Please pick a different user ID</p>)
-            }
+                {
+                    userIDUnique ? (
+                        <div>
+                            <p>User ID is available</p>
+                            <button type="submit">Sign Up</button>
+                        </div>
+                    )
+                        : (<p>User ID already in use. Please pick a different user ID</p>)
+                }
+            </form>
 
             <div>
                 <Link to="/">Return to Login</Link>
